leave_automation v2.0/requirements.txt
//requirements.txt

Flask == 2.3.2
requests == 2.31.0
python-dotenv == 1.0.0
gspread == 5.7.0
oauth2client == 4.1.3
openai == 2.8.1   # or the official client you use; adjust version if needed
